Added auth checking on delete event, changed from requestbody to modelattribute for events so form-data can be mapped

+++ server/src/main/java/calendar/event/EventController.java
@@ -23,8 +23,8 @@ public class EventController {
     // TODO: Update class diagram
     @Autowired
     private EventDAO dao;
-
     private TokenValidator tokenValidator = new TokenValidator();
+    AuthorizationChecker auth = new AuthorizationChecker();
 
     @RequestMapping(value = "/get", method = RequestMethod.GET)
     public List<Event> getEvents(@RequestParam(required = false) String id,
@@ -119,7 +119,6 @@ public class EventController {
     // TODO: Add to docs and diagrams
     @RequestMapping(value = "/get_manageable", method = RequestMethod.GET)
     public List<Event> getManageable() {
-        AuthorizationChecker auth = new AuthorizationChecker();
         List<String> ids = auth.getAllUserIds();
         List<Event> events = new ArrayList<>();
 
@@ -139,19 +138,23 @@ public class EventController {
     }
 
     @RequestMapping(value = "/create", method = RequestMethod.POST)
-    public Event createEvent(@RequestBody CreateEventDTO createEventDTO) {
+    public Event createEvent(@ModelAttribute CreateEventDTO createEventDTO) {
         Event event = new Event(createEventDTO);
         return dao.createEvent(event);
     }
 
+    // TODO: Update seq diagram
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
-    public void deleteEvent(@RequestBody DeleteEventDTO deleteEventDTO) {
-        // TODO: Add authChecker
-        dao.deleteEvent(deleteEventDTO.getId());
+    public void deleteEvent(@ModelAttribute DeleteEventDTO deleteEventDTO) {
+        Event event = dao.getEvent(new ObjectId(deleteEventDTO.getId())).get(0);
+
+        if(auth.currentUserCanManage(event.getCreatedBy())) {
+            dao.deleteEvent(deleteEventDTO.getId());
+        }
     }
 
     @RequestMapping(value = "/update", method = RequestMethod.POST)
-    public Event updateEvent(@RequestBody UpdateEventDTO updateEventDTO) {
+    public Event updateEvent(@ModelAttribute UpdateEventDTO updateEventDTO) {
         // TODO: Add authChecker
         Event event = new Event(updateEventDTO);
         return dao.updateEvent(event);
